// src/app/generated_services/model/item.ts
/**
 * bin_packing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Item { 
    id?: number;
    w: number;
    l: number;
    h: number;
    quantity?: number;
    weight?: number;
    stackable?: boolean;
    uprightOnly?: boolean;
    itemMaxStackWeight?: number | null;
}